o Modified behavior of _handle routines to reuse hash index rather than pointer   to timer. The big advantage of this approach is the extra argument to   start_handle() and stop_handle() routines need not be thread-local. That   makes these routines much easier to use in threaded regions.
  The required type of this second argument was changed from void ** in C to
  int *. And in Fortran from integer*8 to integer. The old method still works
  but there are warning messages from the compiler. Users should change the
  types.

o GPTL now estimates its full overhead rather than just the incurred overhead
  due to the underlying timer. Printed overhead is broken down into an
  estimate of overhead incurred in the timer itself, as well as its parent.

  Also, the timing.* files now contain a breakdown of estimated overhead due
  to hash table lookup, Fortran overhead, finding the entry after hash table
  lookup, and the underlying timing routine.

o Moved as many routines as possible out of gptl.c and into new files
  getoverhead.c, hashstats.c, memstats.c, and pr_summary.c. The size of
  gptl.c was becoming too large for comfort. The change required in some
  cases passing the address of private functions to routines in the new
  files.

+++ ftests/overhead.F90
@@ -8,7 +8,7 @@ program overhead
   real*8 :: wall2, usr2, sys2
   integer :: cycles1, cycles2, cps
   character(len=16) :: msg
-  integer(8) :: handle
+  integer :: handle
 #ifdef THREADED_OMP
   integer, parameter :: maxthreads = 8
 #else
@@ -155,7 +155,7 @@ subroutine do_nothing1 (string, handle)
   implicit none
 
   character(len=*), intent(in) :: string
-  integer(8), intent(in) :: handle
+  integer, intent(in) :: handle
 
   if (string(1:1) == 'x') then
     write(6,*)'Bad string value'
@@ -166,7 +166,7 @@ subroutine do_nothing2 (string, handle)
   implicit none
 
   character(len=*), intent(in) :: string
-  integer(8), intent(in) :: handle
+  integer, intent(in) :: handle
 
   if (string(1:1) == 'x') then
     write(6,*)'Bad string value'
